requirements.txt
django==5.0.6
django-admin-list-filter-dropdown==1.0.3
django-environ==0.11.2
django-taggit==5.0.1
psycopg2-binary==2.9.9
requests==2.32.3

# Dev-only
django-debug-toolbar==4.4.5
